requirements.txt
fastapi[standard]>=0.115.0
uvicorn>=0.30.0
streamlit>=1.40.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
faiss-cpu>=1.9.0
openai>=1.50.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
httpx>=0.27.0
numpy>=1.26.0
